Added ChangeMapAction and Map movement commands. Need hooks to be able to actually move the GameEntity however.

+++ ObjectFramework/RL_Actors/SimpleActorCombat.java
@@ -14,7 +14,7 @@ public class SimpleActorCombat implements CombatComponent {
 
     private SimpleActorCombat() {}
 
-    public SimpleActorCombat(GameEntity owner, int currentHP, int currentArmor, int magicResist) {
+    public SimpleActorCombat(GameEntity owner, int currentHP, int currentArmor, double magicResist) {
 
         this.baseHP = currentHP;
         this.currentHP = currentHP;
